caicon2/requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
einops==0.8.0
filelock==3.14.0
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
huggingface-hub==0.24.6
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.4.0
importlib_resources==6.4.4
Jinja2 @ file:///croot/jinja2_1716993405101/work
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
lap==0.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mat4py==0.6.0
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==2.1.0
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.4
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
motmetrics==1.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
numpy==1.23.5
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.2
psutil==6.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2024.7.24
requests==2.32.3
rich==13.4.2
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.2
trackeval @ git+https://github.com/JonathonLuiten/TrackEval.git@12c8791b303e0a0b50f753af204249e622d0281a
transformers==4.44.2
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ultralytics==8.2.90
ultralytics-thop==2.0.6
urllib3==1.26.20
xmltodict==0.13.0
yapf==0.40.2
zipp==3.20.1
